Graphics works! (Sort of...)
Who doesn't check the width of their variables?
Who makes addresses on a system with a 16 bit address bus 8 bit?
What kind of system only has 256 bytes of memory?

I've done all of these things, but it has been fixed now.......
The display is still mirrored horizontally, but at least the right tiles are put in the right places relative to that....